Updated Makefile
I made a new, proper Makefile. I don't know if I implemented it
properly or not, but it looks clean enough to me lol. I want to
figure out how to change the build directory, because it looks
like a spectre meltdown happened in my code folder after building
all the object files.
On a side note, Dael by Autechre is a glitchy, catchy, and industrail
MASTERPIECE of an IDM song. I think that Miriam might not be into me,
since she hasn't replied to my half funny and half desparate looking
texts after two weeks. Not quite sure on that one, though.